// frontend/src/app/api/models/facetas-destinos.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Continente } from '../models/continente';
import { NivelEscala } from '../models/nivel-escala';
import { RefTaxonomia } from '../models/ref-taxonomia';
import { Slug } from '../models/slug';
import { TramoDuracion } from '../models/tramo-duracion';
export interface FacetasDestinos {
  dificultad: Array<NivelEscala>;
  presupuesto: Array<NivelEscala>;
  regiones: Array<{
'slug': Slug;
'nombre': string;
'continente': Continente;
'paises': Array<RefTaxonomia>;
}>;
  tipos: Array<RefTaxonomia>;
  tramos_duracion: Array<{
'codigo': TramoDuracion;
'etiqueta': string;
'min_dias': number;
'max_dias'?: (number | null);
}>;
}